compressor: make tracked source size an unsigned long long
This value is fed into the zstandard API, which accepts an
unsigned long long.

+++ c-ext/python-zstandard.h
@@ -129,7 +129,7 @@ typedef struct {
 
 	ZstdCompressor* compressor;
 	PyObject* writer;
-	Py_ssize_t sourceSize;
+	unsigned long long sourceSize;
 	size_t outSize;
 	int entered;
 	unsigned long long bytesCompressed;
@@ -162,7 +162,7 @@ typedef struct {
 	ZstdCompressor* compressor;
 	PyObject* reader;
 	Py_buffer buffer;
-	Py_ssize_t sourceSize;
+	unsigned long long sourceSize;
 	size_t readSize;
 
 	int entered;
